Added dropdown menu from the profile icon and fixed hover icons
Added links Log In and Sign Up to dropdown when the user hovers over the profile icon. Once the user logs in, the sign up link will change to the profile link and log in will change to log out. I have also managed to add the hover icons when the user hover overs the nav icons.